Tyre_addGroundContact

+++ acplugin/CustomPhysics/Turbo.cpp
@@ -7,7 +7,7 @@ void Turbo_step(Turbo* pThis, float gas, float rpms, float dt)
 
 	if (rpms > 0.0f && gas > 0.0f)
 	{
-		float fRef = (float)(gas * rpms) / pThis->data.rpmRef;
+		float fRef = (gas * rpms) / pThis->data.rpmRef;
 		fRef = tclamp(fRef, 0.0f, 1.0f);
 		fGamma = powf(fRef, pThis->data.gamma);
 	}
@@ -26,15 +26,14 @@ void Turbo_step(Turbo* pThis, float gas, float rpms, float dt)
 		fLag = tclamp(fLag, 0.0f, 1.0f);
 	}
 
-	float fFinalRotation = (float)((float)(fGamma - fRotation) * fLag) + fRotation;
+	float fFinalRotation = ((fGamma - fRotation) * fLag) + fRotation;
 	pThis->rotation = fFinalRotation;
 
 	float fWastegate = pThis->data.wastegate;
-
 	if (fWastegate != 0.0f)
 	{
 		float fUserWG = fWastegate * pThis->userSetting;
-		if ((float)(pThis->data.maxBoost * pThis->rotation) > fUserWG)
+		if ((pThis->data.maxBoost * pThis->rotation) > fUserWG)
 		{
 			pThis->rotation = fUserWG / pThis->data.maxBoost;
 		}
